--- a/src/include/utils/acl.h
+++ b/src/include/utils/acl.h
@@ -333,4 +333,7 @@ extern bool has_bypassrls_privilege(Oid roleid);
 typedef bool (*tsql_has_linked_srv_permissions_hook_type) (Oid roleid);
 extern PGDLLIMPORT tsql_has_linked_srv_permissions_hook_type tsql_has_linked_srv_permissions_hook;
 
+typedef Oid (*bbf_get_sysadmin_oid_hook_type) (void);
+extern PGDLLIMPORT bbf_get_sysadmin_oid_hook_type bbf_get_sysadmin_oid_hook;
+
 #endif							/* ACL_H */